Move optimizer to sigma.parser.condition to enable it for all backends
